% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dict_path}
\alias{dict_path}
\title{Get the path to a dictionary file}
\usage{
dict_path(
  dict = c("ipadic-mecab-2_7_0", "jumandic-mecab-7_0", "naist-jdic-mecab-0_6_3b",
    "unidic-mecab-2_1_2", "unidic-cwj-3_1_1"),
  dict_dir = NULL
)
}
\arguments{
\item{dict}{Dictionary name.}

\item{dict_dir}{Directory where dictionaries are placed.
By default, the return value \code{rappdirs::user_cache_dir("vibrrt")} is used.}
}
\description{
Get the path to a dictionary file
}
