% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{download_dict}
\alias{download_dict}
\title{Download dictionary file}
\usage{
download_dict(
  dict = c("ipadic-mecab-2_7_0", "jumandic-mecab-7_0", "naist-jdic-mecab-0_6_3b",
    "unidic-mecab-2_1_2", "unidic-cwj-3_1_1+compact-dual"),
  dict_dir = NULL
)
}
\arguments{
\item{dict}{Dictionary name.}

\item{dict_dir}{Directory where dictionaries are placed.
By default, the return value \code{rappdirs::user_cache_dir("vibrrt")} is used.}
}
\value{
The path to the 'system.dic' file is returned invisibly.
}
\description{
Download dictionary file
}
\seealso{
\url{https://github.com/daac-tools/vibrato/releases/tag/v0.5.0}
}
