% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{tokenize}
\alias{tokenize}
\title{Tokenize sentence for data.frame}
\usage{
tokenize(
  tbl,
  text_field = "text",
  docid_field = "doc_id",
  sys_dic = system.file("dict/ipadic-mecab-2_7_0/system.dic", package = "vibrrt"),
  split = FALSE
)
}
\arguments{
\item{tbl}{A data.frame.}

\item{text_field}{String or symbol; column name where to get texts to be tokenized.}

\item{docid_field}{String or symbol; column name where to get identifiers of texts.}

\item{sys_dic}{Character scalar; path to the system dictionary for 'vibrato'.}

\item{split}{Logical. If supplied `TRUE`, the function internally splits the sentence
into sub-sentences using \code{stringi::stri_split_boudaries(type = "sentence")}.}
}
\value{
data.frame.
}
\description{
Tokenize sentence for data.frame
}
