% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{tokenize_vec}
\alias{tokenize_vec}
\title{Tokenize sentence for character vector}
\usage{
tokenize_vec(
  sentence,
  sys_dic = system.file("dict/ipadic-mecab-2_7_0/system.dic", package = "vibrrt"),
  user_dic = "",
  split = FALSE,
  mode = c("parse", "wakati")
)
}
\arguments{
\item{sentence}{Character vector to be tokenized.}

\item{sys_dic}{Character scalar; path to the system dictionary for 'vibrato'.}

\item{user_dic}{Character scalar; path to the user dictionary for 'vibrato'.}

\item{split}{Logical. If supplied `TRUE`, the function internally splits the sentence
into sub-sentences using \code{stringi::stri_split_boudaries(type = "sentence")}.}

\item{mode}{Character scalar to switch output format.}
}
\value{
data.frame or named list.
}
\description{
Tokenize sentence for character vector
}
