% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{tokenize}
\alias{tokenize}
\title{Tokenize sentences using a tagger}
\usage{
tokenize(
  x,
  text_field = "text",
  docid_field = "doc_id",
  split = FALSE,
  mode = c("parse", "wakati"),
  tagger
)
}
\arguments{
\item{x}{A data.frame like object or a character vector to be tokenized.}

\item{text_field}{<\code{\link[rlang:args_data_masking]{data-masked}}>
String or symbol; column containing texts to be tokenized.}

\item{docid_field}{<\code{\link[rlang:args_data_masking]{data-masked}}>
String or symbol; column containing document IDs.}

\item{split}{split Logical. When passed as \code{TRUE}, the function
internally splits the sentences into sub-sentences}

\item{mode}{Character scalar to switch output format.}

\item{tagger}{A tagger function created by \code{\link[=create_tagger]{create_tagger()}}.}
}
\value{
A tibble or a named list of tokens.
}
\description{
Tokenize sentences using a tagger
}
