% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{vibrrt-package}
\alias{vibrrt}
\alias{vibrrt-package}
\title{vibrrt: An R Wrapper for 'vibrato'}
\description{
An R wrapper for 'vibrato' (\url{https://github.com/daac-tools/vibrato}), Viterbi-based accelerated tokenizer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paithiov909.github.io/vibrrt/}
  \item Report bugs at \url{https://github.com/paithiov909/vibrrt/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
